pyproject.toml
[project]
name = "llama-index-onboarding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "ipykernel>=7.1.0",
    "llama-cloud-services>=0.6.85",
    "llama-index-core>=0.14.9",
    "llama-index-llms-openai>=0.6.10",
    "llama-index-workflows>=2.11.5",
]
